transparent background image
On the themes that already use a background, the output will look the exact same. (both white and inverted)
However, this file can be used with other background colors, too!
This commit will not make any visible change for the user yet, but provides for easier addition of background-changing themes.